Move constants which are using import.meta.env to separate file to avoid errors in TestCafe tests
