+++ src/domain/keyword/utils.ts
@@ -2,7 +2,8 @@ import { ApolloClient, NormalizedCacheObject } from '@apollo/client';
 import { TFunction } from 'i18next';
 
 import { MenuItemOptionProps } from '../../common/components/menuDropdown/types';
-import { LINKED_EVENTS_SYSTEM_DATA_SOURCE, ROUTES } from '../../constants';
+import { ROUTES } from '../../constants';
+import { LINKED_EVENTS_SYSTEM_DATA_SOURCE } from '../../envVariables';
 import {
   CreateKeywordMutationInput,
   Keyword,
