+++ src/domain/keywordSet/utils.ts
@@ -2,7 +2,8 @@ import { TFunction } from 'i18next';
 import capitalize from 'lodash/capitalize';
 
 import { MenuItemOptionProps } from '../../common/components/menuDropdown/types';
-import { LINKED_EVENTS_SYSTEM_DATA_SOURCE, ROUTES } from '../../constants';
+import { ROUTES } from '../../constants';
+import { LINKED_EVENTS_SYSTEM_DATA_SOURCE } from '../../envVariables';
 import {
   CreateKeywordSetMutationInput,
   KeywordFieldsFragment,
