+++ src/domain/organization/utils.ts
@@ -2,11 +2,8 @@ import { ApolloClient, NormalizedCacheObject } from '@apollo/client';
 import { TFunction } from 'i18next';
 
 import { MenuItemOptionProps } from '../../common/components/menuDropdown/types';
-import {
-  DATE_FORMAT_API,
-  LINKED_EVENTS_SYSTEM_DATA_SOURCE,
-  ROUTES,
-} from '../../constants';
+import { DATE_FORMAT_API, ROUTES } from '../../constants';
+import { LINKED_EVENTS_SYSTEM_DATA_SOURCE } from '../../envVariables';
 import {
   CreateOrganizationMutationInput,
   OrganizationDocument,
